requirements.txt
slack-sdk>=3.27.0
python-dotenv>=1.0.0
PyYAML>=6.0
langchain-text-splitters>=0.2.0
tiktoken>=0.7.0
